Add missing file to fix unit tests
