+++ pkg/util.go
@@ -19,11 +19,6 @@ package validator
 import (
 	"context"
 
-	"github.com/ipfs/go-cid"
-	blockstore "github.com/ipfs/go-ipfs-blockstore"
-	dshelp "github.com/ipfs/go-ipfs-ds-help"
-	"github.com/jmoiron/sqlx"
-
 	"github.com/ipfs/go-blockservice"
 	"github.com/ipfs/kubo/core"
 	"github.com/ipfs/kubo/repo/fsrepo"
@@ -45,37 +40,4 @@ func InitIPFSBlockService(ipfsPath string) (blockservice.BlockService, error) {
 		return nil, newNodeErr
 	}
 	return ipfsNode.Blocks, nil
-}
-
-// PublishRaw derives a cid from raw bytes and provided codec and multihash type, and writes it to the db tx
-func PublishRaw(tx *sqlx.Tx, codec, mh uint64, raw []byte, blockNumber uint64) (string, error) {
-	c, err := RawdataToCid(codec, raw, mh)
-	if err != nil {
-		return "", err
-	}
-	dbKey := dshelp.MultihashToDsKey(c.Hash())
-	prefixedKey := blockstore.BlockPrefix.String() + dbKey.String()
-	_, err = tx.Exec(`INSERT INTO public.blocks (key, data, block_number) VALUES ($1, $2, $3) ON CONFLICT DO NOTHING`, prefixedKey, raw, blockNumber)
-	return c.String(), err
-}
-
-// RawdataToCid takes the desired codec, multihash type, and a slice of bytes
-// and returns the proper cid of the object.
-func RawdataToCid(codec uint64, rawdata []byte, multiHash uint64) (cid.Cid, error) {
-	c, err := cid.Prefix{
-		Codec:    codec,
-		Version:  1,
-		MhType:   multiHash,
-		MhLength: -1,
-	}.Sum(rawdata)
-	if err != nil {
-		return cid.Cid{}, err
-	}
-	return c, nil
-}
-
-// ResetTestDB drops all rows in the test db public.blocks table
-func ResetTestDB(db *sqlx.DB) error {
-	_, err := db.Exec("DELETE FROM public.blocks")
-	return err
 }
